Move files to correct directory
