+++ src/main/java/com/thebois/controllers/gamecontrollers/StructureController.java
@@ -7,7 +7,7 @@ import com.thebois.controllers.AbstractInputProcessor;
 import com.thebois.controllers.IController;
 import com.thebois.models.world.World;
 import com.thebois.views.IProjector;
-import com.thebois.views.IView;
+import com.thebois.views.gameviews.IView;
 import com.thebois.views.gameviews.StructureView;
 
 /**
