+++ src/main/java/com/thebois/views/gameviews/StructureView.java
@@ -5,7 +5,6 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 
 import com.thebois.models.world.structures.IStructure;
-import com.thebois.views.IView;
 import com.thebois.views.TextureUtils;
 
 /**
